Fix review comments, Iter. 6 -- remove MetadataBlockRepositoryExt
